refactor: Car 생성시 예외처리 에러메시지 별도 상수로 관리
