+++ __tests__/Car.test.js
@@ -1,4 +1,8 @@
-import { Car } from '../src/Car.js';
+import { Car } from '../src/models/Car/index.js';
+import {
+  CAR_NAME_REQUIRED_ERROR_MESSAGE,
+  CAR_NAME_MAX_LENGTH_ERROR_MESSAGE,
+} from '../src/models/Car/constant.js';
 
 const createCar = (name = '이름') => {
   return new Car(name);
@@ -14,16 +18,14 @@ describe('자동차는', () => {
   });
 
   it('이름이 빈 문자열이거나 공백일 경우 에러가 발생한다.', () => {
-    const ERROR_MESSAGE = '자동차 이름은 필수입니다.';
-
-    expect(() => createCar('')).toThrow(ERROR_MESSAGE);
-    expect(() => createCar('   ')).toThrow(ERROR_MESSAGE);
+    expect(() => createCar('')).toThrow(CAR_NAME_REQUIRED_ERROR_MESSAGE);
+    expect(() => createCar('   ')).toThrow(CAR_NAME_REQUIRED_ERROR_MESSAGE);
   });
 
   it('이름이 5자를 넘어가면 에러가 발생한다.', () => {
-    const ERROR_MESSAGE = '자동차 이름은 5자 이하이어야 합니다.';
-
-    expect(() => createCar('이름이6글자')).toThrow(ERROR_MESSAGE);
+    expect(() => createCar('이름이6글자')).toThrow(
+      CAR_NAME_MAX_LENGTH_ERROR_MESSAGE,
+    );
   });
 
   it('위치 값을 가지며, 초기 상태는 0 이다', () => {
