+++ src/models/Car/constant.js
@@ -1,0 +1,4 @@
+export const CAR_NAME_MAX_LENGTH = 5;
+
+export const CAR_NAME_REQUIRED_ERROR_MESSAGE = '자동차 이름은 필수입니다.';
+export const CAR_NAME_MAX_LENGTH_ERROR_MESSAGE = `자동차 이름은 ${CAR_NAME_MAX_LENGTH}자 이하이어야 합니다.`;
