+++ src/models/Car/index.js
@@ -1,0 +1,23 @@
+import {
+  CAR_NAME_MAX_LENGTH,
+  CAR_NAME_REQUIRED_ERROR_MESSAGE,
+  CAR_NAME_MAX_LENGTH_ERROR_MESSAGE,
+} from './constant.js';
+
+export class Car {
+  constructor(name) {
+    if (!name || !name.trim()) {
+      throw new Error(CAR_NAME_REQUIRED_ERROR_MESSAGE);
+    }
+    if (name.trim().length > CAR_NAME_MAX_LENGTH) {
+      throw new Error(CAR_NAME_MAX_LENGTH_ERROR_MESSAGE);
+    }
+
+    this.name = name.trim();
+    this.location = 0;
+  }
+
+  move() {
+    this.location += 1;
+  }
+}
